goals/requirements.txt
Flask==2.3.2
pymongo==4.2.0
gunicorn==20.1.0
python-dotenv==0.19.0
Flask-PyMongo==2.3.0
Flask-CORS==3.0.10
pytest==6.2.5
werkzeug==2.3.3